requirements.txt
wxPython
google-genai
keyboard
pywin32
